shared restore from backup function
