+++ tests/unit/test_nfo_utils.py
@@ -4,9 +4,7 @@ import tempfile
 from pathlib import Path
 
 from sonarr_metadata_rewrite.nfo_utils import (
-    create_backup,
     find_target_files,
-    get_backup_path,
     is_nfo_file,
     is_rewritable_image,
 )
@@ -280,139 +278,3 @@ class TestFindRewritableImages:
                 p for p in find_target_files(temp_path) if is_rewritable_image(p)
             ]
             assert found_files == []
-
-
-class TestBackupFunctions:
-    """Test create_backup and get_backup_path functions."""
-
-    def test_backup_with_none_backup_dir(self) -> None:
-        """Test backup functions with None backup_dir."""
-        with tempfile.TemporaryDirectory() as temp_dir:
-            temp_path = Path(temp_dir)
-            file_path = temp_path / "test.nfo"
-            file_path.write_text("<tvshow></tvshow>")
-
-            # create_backup returns False
-            result = create_backup(file_path, None, temp_path)
-            assert result is False
-
-            # get_backup_path returns None
-            backup_path = get_backup_path(file_path, None, temp_path)
-            assert backup_path is None
-
-    def test_backup_nonexistent_file(self) -> None:
-        """Test that create_backup returns False for nonexistent file."""
-        with tempfile.TemporaryDirectory() as temp_dir:
-            temp_path = Path(temp_dir)
-            backup_dir = temp_path / "backup"
-            file_path = temp_path / "nonexistent.nfo"
-
-            result = create_backup(file_path, backup_dir, temp_path)
-            assert result is False
-
-    def test_backup_and_retrieval_workflow(self) -> None:
-        """Test complete workflow: create backup and retrieve it."""
-        with tempfile.TemporaryDirectory() as temp_dir:
-            temp_path = Path(temp_dir)
-            backup_dir = temp_path / "backup"
-            file_path = temp_path / "test.nfo"
-            content = "<tvshow></tvshow>"
-            file_path.write_text(content)
-
-            # Before backup, get_backup_path returns None
-            backup_path = get_backup_path(file_path, backup_dir, temp_path)
-            assert backup_path is None
-
-            # Create backup
-            created = create_backup(file_path, backup_dir, temp_path)
-            assert created is True
-
-            # After backup, get_backup_path returns the backup path
-            backup_path = get_backup_path(file_path, backup_dir, temp_path)
-            assert backup_path is not None
-            assert backup_path.exists()
-            assert backup_path.read_text() == content
-
-    def test_backup_preserves_directory_structure(self) -> None:
-        """Test backup maintains directory structure."""
-        with tempfile.TemporaryDirectory() as temp_dir:
-            temp_path = Path(temp_dir)
-            backup_dir = temp_path / "backup"
-            subdir = temp_path / "shows" / "series1"
-            subdir.mkdir(parents=True)
-            file_path = subdir / "tvshow.nfo"
-            content = "<tvshow></tvshow>"
-            file_path.write_text(content)
-
-            # Create backup
-            result = create_backup(file_path, backup_dir, temp_path)
-            assert result is True
-
-            # Verify structure is preserved
-            expected_backup = backup_dir / "shows" / "series1" / "tvshow.nfo"
-            assert expected_backup.exists()
-            assert expected_backup.read_text() == content
-
-            # get_backup_path should find it
-            backup_path = get_backup_path(file_path, backup_dir, temp_path)
-            assert backup_path == expected_backup
-
-    def test_backup_does_not_overwrite_existing(self) -> None:
-        """Test that backup doesn't overwrite existing backup."""
-        with tempfile.TemporaryDirectory() as temp_dir:
-            temp_path = Path(temp_dir)
-            backup_dir = temp_path / "backup"
-            file_path = temp_path / "test.nfo"
-            file_path.write_text("<tvshow>new</tvshow>")
-
-            # Create existing backup with different content
-            backup_path = backup_dir / "test.nfo"
-            backup_path.parent.mkdir(parents=True)
-            original_content = "<tvshow>original</tvshow>"
-            backup_path.write_text(original_content)
-
-            # Try to backup new content
-            result = create_backup(file_path, backup_dir, temp_path)
-            assert result is True
-
-            # Verify backup wasn't overwritten
-            assert backup_path.read_text() == original_content
-
-            # get_backup_path should return existing backup
-            retrieved_path = get_backup_path(file_path, backup_dir, temp_path)
-            assert retrieved_path == backup_path
-            assert retrieved_path is not None
-            assert retrieved_path.read_text() == original_content
-
-    def test_backup_stem_matching_for_different_extensions(self) -> None:
-        """Test both functions handle same stem with different extensions."""
-        with tempfile.TemporaryDirectory() as temp_dir:
-            temp_path = Path(temp_dir)
-            backup_dir = temp_path / "backup"
-
-            # Create original poster.png in backup
-            backup_path_png = backup_dir / "poster.png"
-            backup_path_png.parent.mkdir(parents=True)
-            backup_path_png.write_bytes(b"PNG original")
-
-            # Try to backup poster.jpg (same stem, different extension)
-            file_path_jpg = temp_path / "poster.jpg"
-            file_path_jpg.write_bytes(b"JPG new")
-
-            # create_backup should recognize existing stem and not create new
-            result = create_backup(file_path_jpg, backup_dir, temp_path)
-            assert result is True
-
-            # Verify original backup still exists and wasn't modified
-            assert backup_path_png.exists()
-            assert backup_path_png.read_bytes() == b"PNG original"
-
-            # Verify new backup wasn't created
-            backup_path_jpg = backup_dir / "poster.jpg"
-            assert not backup_path_jpg.exists()
-
-            # get_backup_path should find the .png backup when looking for .jpg
-            retrieved_path = get_backup_path(file_path_jpg, backup_dir, temp_path)
-            assert retrieved_path == backup_path_png
-            assert retrieved_path is not None
-            assert retrieved_path.read_bytes() == b"PNG original"
